requirements.txt
gsheetsdb==0.1.13.1
pandas==2.0.3
Pillow==10.0.1
protobuf==3.12.4
streamlit==1.27.1
